+++ customer/api_test.go
@@ -40,7 +40,7 @@ var mockedResponses = map[string]string{
 		"chat_id": "PJ0MRSHTDG",
 		"thread_id": "PGDGHT5G"
 	}`,
-	"activate_chat": `{
+	"resume_chat": `{
 		"thread_id": "PGDGHT5G"
 	}`,
 	"send_event": `{
@@ -285,7 +285,7 @@ func createMockedResponder(t *testing.T, method string) roundTripFunc {
 			}
 		}
 
-		if req.URL.String() != "https://api.livechatinc.com/v3.2/customer/action/"+method+"?license_id=12345" {
+		if req.URL.String() != "https://api.livechatinc.com/v3.3/customer/action/"+method+"?license_id=12345" {
 			t.Errorf("Invalid URL for Customer API request: %s", req.URL.String())
 			return createServerError("Invalid URL")
 		}
@@ -375,7 +375,7 @@ func TestStartChatShouldReturnDataReceivedFromCustomerAPI(t *testing.T) {
 			Events: []interface{}{m},
 		},
 	}
-	chatID, threadID, _, rErr := api.StartChat(ic, true)
+	chatID, threadID, _, rErr := api.StartChat(ic, true, true)
 	if rErr != nil {
 		t.Errorf("StartChat failed: %v", rErr)
 	}
@@ -446,17 +446,17 @@ func TestSendSystemMessageShouldReturnDataReceivedFromCustomerAPI(t *testing.T) 
 	}
 }
 
-func TestActivateChatShouldReturnDataReceivedFromCustomerAPI(t *testing.T) {
-	client := NewTestClient(createMockedResponder(t, "activate_chat"))
+func TestResumeChatShouldReturnDataReceivedFromCustomerAPI(t *testing.T) {
+	client := NewTestClient(createMockedResponder(t, "resume_chat"))
 
 	api, err := customer.NewAPI(stubTokenGetter, client, "client_id")
 	if err != nil {
 		t.Errorf("API creation failed")
 	}
 
-	threadID, _, rErr := api.ActivateChat(&objects.InitialChat{}, true)
+	threadID, _, rErr := api.ResumeChat(&objects.InitialChat{}, true, true)
 	if rErr != nil {
-		t.Errorf("ActivateChat failed: %v", rErr)
+		t.Errorf("ResumeChat failed: %v", rErr)
 	}
 
 	if threadID != "PGDGHT5G" {
@@ -884,7 +884,7 @@ func TestStartChatShouldNotCrashOnErrorResponse(t *testing.T) {
 		t.Errorf("API creation failed")
 	}
 
-	_, _, _, rErr := api.StartChat(&objects.InitialChat{}, true)
+	_, _, _, rErr := api.StartChat(&objects.InitialChat{}, true, true)
 	verifyErrorResponse("StartChat", rErr, t)
 }
 
@@ -900,16 +900,16 @@ func TestSendEventShouldNotCrashOnErrorResponse(t *testing.T) {
 	verifyErrorResponse("SendEvent", rErr, t)
 }
 
-func TestActivateChatShouldNotCrashOnErrorResponse(t *testing.T) {
-	client := NewTestClient(createMockedErrorResponder(t, "activate_chat"))
+func TestResumeChatShouldNotCrashOnErrorResponse(t *testing.T) {
+	client := NewTestClient(createMockedErrorResponder(t, "resume_chat"))
 
 	api, err := customer.NewAPI(stubTokenGetter, client, "client_id")
 	if err != nil {
 		t.Errorf("API creation failed")
 	}
 
-	_, _, rErr := api.ActivateChat(&objects.InitialChat{}, true)
-	verifyErrorResponse("ActivateChat", rErr, t)
+	_, _, rErr := api.ResumeChat(&objects.InitialChat{}, true, true)
+	verifyErrorResponse("ResumeChat", rErr, t)
 }
 
 func TestListChatsShouldNotCrashOnErrorResponse(t *testing.T) {